# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Management API",
        "description": "API documentation for managing hotel rooms, guests, and reservations",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get the list of guests",
                "operationId": "e7585a0ec4b0672b62f3e98d2514d54c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of guests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Guest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Create a new guest",
                "operationId": "448b4d4333d687f0f6111e6814d0d8cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Guest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get a specific guest by ID",
                "operationId": "5c06d71761da36a9db7936ee6fe976d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single guest",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update a guest",
                "operationId": "117eaa66abd209a6833cc2959caeff85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Guest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Delete a guest",
                "operationId": "153a4fa10e9e7c8bc84e46b3885b6e47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guest deleted successfully"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get the list of rooms",
                "operationId": "f36201d4142b0b6e142bcd5e5d0bcc1e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of rooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Room"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "5926474104d331e075299a5b9d2f7416",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Room"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a specific room by ID",
                "operationId": "2fb25667a180bad0871584c60fa392fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single room",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "operationId": "a16e22295fb7ccec81d4a685ab6edf52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Room"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "9f216809b68e2ce14c857a26768e9228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Guest": {
                "required": [
                    "first_name",
                    "last_name"
                ],
                "properties": {
                    "id": {
                        "description": "Guest ID",
                        "type": "integer"
                    },
                    "first_name": {
                        "description": "First name of the guest",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name of the guest",
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Room": {
                "required": [
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Room ID",
                        "type": "integer"
                    },
                    "room_number": {
                        "description": "Room number",
                        "type": "string"
                    },
                    "status": {
                        "description": "Room status",
                        "type": "string",
                        "enum": [
                            "ready",
                            "pending_cleanup",
                            "reserved"
                        ]
                    },
                    "created_at": {
                        "description": "The date and time when the room was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the room was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Guests",
            "description": "Operations related to guests"
        },
        {
            "name": "Rooms",
            "description": "Operations related to rooms"
        }
    ]
}